% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter-pkg.R, R/converter.R
\docType{package}
\name{converter}
\alias{converter}
\alias{converter-package}
\title{Converter}
\usage{
converter(x, to)
}
\arguments{
\item{x}{numeric vector}

\item{to}{string value; "sqkm" converts ha to sqkm and "ha" convert sqkm to ha}
}
\value{
numeric vector
}
\description{
It allows to convert area units from square kilometers to hectares and viceversa
}
\examples{
x<-(300:310)
converter(x,"sqkm")
converter(x,"ha")

}
